Fix integer overflow when loading time exceeds uint16 range
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@11900 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/descriptor/vehicle_desc.h
@@ -238,7 +238,7 @@ public:
 	bool can_follow_any() const { return trailer_count==0; }
 
 	uint16 get_capacity() const { return capacity; }
-	uint16 get_loading_time() const { return loading_time; } // ms per full loading/unloading
+	uint32 get_loading_time() const { return loading_time; } // ms per full loading/unloading
 	uint32 get_weight() const { return weight; }
 	uint32 get_power() const { return power; }
 	sint64 get_running_cost() const { return running_cost; }
